Fixed up some eslint warnings, moved styles to their own file.
